create a index.desktop.js and index.website.js file
